Use spring mvc to configure a restful server.

+++ src/main/java/com/tusk/lvoryTower/dao/UserDAO.java
@@ -4,5 +4,5 @@ import com.tusk.lvoryTower.module.User;
 
 public interface UserDAO {
 	public int countUser();
-	public User getUserById(int userId);
+	public User getUserById(long userId);
 }
