+++ src/main/java/com/tusk/lvoryTower/services/UserService.java
@@ -4,5 +4,5 @@ import com.tusk.lvoryTower.module.User;
 
 public interface UserService {
 	public int getUserCount();
-	public User getUserById(int userId);
+	public User getUserById(long userId);
 }
